data: add new sources
