+++ back/iarbre_data/tests/test_c03_data.py
@@ -1,10 +1,10 @@
 from django.test import TestCase
 from iarbre_data.data_config import DATA_FILES, URL_FILES
 from iarbre_data.models import Data
+from iarbre_data.management.commands.utils_data import apply_actions
 from iarbre_data.management.commands.c03_import_data import (
     download_from_url,
     read_data,
-    apply_actions,
     process_data,
     save_geometries,
 )
